added new page for sales upload, split it off of sales list page
